Hotfix: fix bug with uint16 images https://github.com/NVIDIA-ISAAC-ROS/isaac_ros_nvblox/issues/10 and https://github.com/NVIDIA-ISAAC-ROS/isaac_ros_nvblox/issues/9
