+++ nvblox_ros/package.xml
@@ -35,7 +35,9 @@ license agreement from NVIDIA CORPORATION is strictly prohibited.
   <depend>std_srvs</depend>
   <depend>tf2_eigen</depend>
   <depend>tf2_ros</depend>
-  
+  <depend>message_filters</depend>
+  <depend>nvblox</depend>
+
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
